Change slide to linear heights, and do an abstraction on OpModes
This commit changes the slide so that specific button presses change the slide to junction heights, and heights of cones. These values still need to be adjusted.
Additionally, code for the slide and claw was moved to UpperSystemManager. This class has the same functions as before, but is instead contained in one class.

AssertingDominance.java needs to be fixed due to slide changes.